% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_map.R
\name{best_map2}
\alias{best_map2}
\title{Title}
\usage{
best_map2(
  .x,
  .y,
  .f,
  ...,
  .workers = NULL,
  .scheduling = 1,
  .silent = FALSE,
  .show_progress = TRUE,
  .globals = TRUE,
  .pkg = NULL
)
}
\arguments{
\item{.x}{...}

\item{.y}{...}

\item{.f}{...}

\item{...}{...}

\item{.workers}{...}

\item{.scheduling}{...}

\item{.silent}{...}

\item{.show_progress}{...}

\item{.globals}{...}

\item{.pkg}{...}
}
\value{
...
}
\description{
Title
}
\examples{
"example"
}
