% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_df.R
\name{chunk_df}
\alias{chunk_df}
\title{Title}
\usage{
chunk_df(.df, .n_rows = NULL, .n_chunks = NULL)
}
\arguments{
\item{.df}{...}

\item{.n_rows}{...}

\item{.n_chunks}{...}
}
\value{
...
}
\description{
Title
}
\examples{
"example"
}
