% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_df.R
\name{chunk_vector}
\alias{chunk_vector}
\title{Title}
\usage{
chunk_vector(.vec, .n_elems = NULL, .n_chunks = NULL)
}
\arguments{
\item{.vec}{...}

\item{.n_elems}{...}

\item{.n_chunks}{...}
}
\value{
...
}
\description{
Title
}
\examples{
"example"
}
